get mindspore supported arch from package
