--- python/sglang/srt/models/mindspore_models/layers/__init__.py
@@ -1,22 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the SGLang project
-from sglang.srt.models.mindspore_models.layers.activation import SwiGLU
-from sglang.srt.models.mindspore_models.layers.attention import MsNativeAttnBackend
-from sglang.srt.models.mindspore_models.layers.linear import (
-    ColParallelLinear,
-    MLPColParallelLinear,
-    MoeReplicatedLinear,
-    QKVParallelLinear,
-    RowParallelLinear,
-)
-from sglang.srt.models.mindspore_models.layers.moe import *
-from sglang.srt.models.mindspore_models.layers.norm import RMSNorm
-from sglang.srt.models.mindspore_models.layers.rope import (
-    BaseRotaryEmbedding,
-    DeepseekScalingRotaryEmbedding,
-    YaRNScalingRotaryEmbedding,
-    yarn_get_mscale,
-)
-from sglang.srt.models.mindspore_models.layers.vocab_embedding import (
-    VocabParallelEmbedding,
-)
