--- python/sglang/srt/models/mindspore_models/layers/moe/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the SGLang project
-from sglang.srt.models.mindspore_models.layers.moe.fused_moe import FusedMoe
